refactored command validation into own function
