+++ src/cluster-handler.ts
@@ -2,55 +2,19 @@ import * as cluster from "cluster";
 import * as tmi from "tmi.js";
 import { ChatCommandFactory } from "./chat-command-factory";
 import { ChatCommandJob } from "./chat-command-job.model";
-import { IChatCommand, IChatCommandJob, IChatCommandResult } from "./chat-command.interface";
-import { DiceCommand } from "./commands/dice.model";
-import { GreetingCommand } from "./commands/greeting.model";
+import { IChatCommand, IChatCommandContext, IChatCommandJob, IChatCommandResult } from "./chat-command.interface";
 import { getTwitchClient, globals, sleep } from "./twitch-client";
-
-const availableCommands: IChatCommand[] = [new GreetingCommand(), new DiceCommand()];
+import { validateCommand } from "./chat-command-validator";
 
 export function processClusterMessage(target: string, sender: tmi.Userstate, msg: string, isSelf: boolean) {
     if (isSelf) return;
-    if (msg.indexOf("!") < 0) {
-        console.log(`* message dropped: no command - message: ${msg}, sender: ${sender.username}`);
+    const context: IChatCommandContext | undefined = validateCommand(target, sender, msg);
+
+    if (context === undefined) {
         return;
     }
+    const job: IChatCommandJob = new ChatCommandJob(context.command, context.sender, context.recipients);
 
-    console.log(`* command received: no command - message: ${msg}, sender: ${sender.username}`);
-
-    const commandsWithReceiver = msg.match(/!\w+(\s*@[\w|\d]*)*/g);
-    if (!commandsWithReceiver || commandsWithReceiver.length <= 0) {
-        console.log(`* command dropped: no command - message: ${msg}, sender: ${sender.username}`);
-        return;
-    }
-
-    const firstCommandWithReceivers = commandsWithReceiver[0].trim().toLowerCase();
-    const matchingAvailableCommand = availableCommands.find(
-        (x) =>
-            (!Array.isArray(x.trigger) && firstCommandWithReceivers.indexOf(x.trigger.toLowerCase()) === 0) ||
-            (Array.isArray(x.trigger) &&
-                x.trigger.some((trigger) => firstCommandWithReceivers.indexOf(trigger.toLowerCase()) === 0)),
-    );
-
-    if (!matchingAvailableCommand) {
-        console.log("* dropping command: no handler found");
-        return;
-    }
-
-    if (
-        !matchingAvailableCommand.permittedUsers?.some((x) => x.toLowerCase() === sender.username?.toLowerCase()) &&
-        (matchingAvailableCommand.allowedForMods || matchingAvailableCommand.allowedForSubscriber) &&
-        (!(matchingAvailableCommand.allowedForMods && sender.mod) ||
-            !(matchingAvailableCommand.allowedForSubscriber && sender.subscriber))
-    ) {
-        console.log(`* dropping command: no permission ${sender.username}: ${msg}`);
-        return;
-    }
-
-    const recipients = firstCommandWithReceivers.match(/@[\w|\d]*/g);
-    const job: IChatCommandJob = new ChatCommandJob(matchingAvailableCommand, sender.username, recipients);
-
-    // matchingAvailableCommand.execute(recipients, sender.username)
     const listener = (worker: cluster.Worker, result: IChatCommandResult) => {
         // this code is executed in context of the master
         if (result.messages !== undefined) {
